removed strategy basics because we moved the rule based strategy to strategies and the enum class to the order.py file in api_getter pack. This is because it makes more sende; strategy is a pack that builds strategies for the api_getterpack where the enum class represents the type of side for the order
