+++ src/strategies/bollinger_bands_strategy.py
@@ -1,7 +1,8 @@
+from ..api_getter.order import SideSignal
+
 import requests
 from datetime import datetime, timedelta, timezone
 from typing import Tuple
-from .strategy_basics import SideSignal
 from config import load_api_keys 
 
 def bollinger_bands_strategy(position: dict) -> Tuple[SideSignal, int]:
