+++ src/strategies/mean_reversion.py
@@ -1,5 +1,4 @@
-
-from .strategy_basics import SideSignal
+from ..api_getter.order import SideSignal
 
 def mean_reversion_strategy(position_data: dict, moving_avg_price: float) -> tuple[SideSignal, int]:
     """
