+++ src/strategies/momentum.py
@@ -1,5 +1,4 @@
-
-from .strategy_basics import SideSignal
+from ..api_getter.order import SideSignal
 
 def momentum_strategy(position_data: dict) -> tuple[SideSignal, int]:
     """
