+++ src/strategies/rsi.py
@@ -1,8 +1,8 @@
+from ..api_getter.order import SideSignal
 
 import requests
 from datetime import datetime, timedelta, timezone
 from typing import Tuple
-from .strategy_basics import SideSignal
 from config import load_api_keys
 
 def calculate_rsi(closes, period: int = 14) -> float:
